refactor(security): transition to interceptor based JWT authentication
Key changes include:
- Auth Interceptor `AuthInterceptor`:
    - Implemented `HandlerInterceptor` to perform JWT extraction, validation, and `Member` lookup before controller execution.
    - Stores the authenticated `Member` object in request attributes for access.
    - Throws `UnauthorizedException` for any authentication failures.
- Simplified Argument Resolver `LoginMemberArgumentResolver`:
    - Modified to only retrieve the `Member` object from request attributes, removing direct JWT processing logic.
    - No longer directly depends on `JwtTokenProvider` or `AuthService`.
- Web Configuration `WebConfig`:
    - Registered `AuthInterceptor` to apply to protected API paths for `/api/cart/**`.
- Auth Service `AuthService`:
    - `getMemberById` method now used by the interceptor to fetch the authenticated user.
