+++ src/main/kotlin/ecommerce/application/WebConfig.kt
@@ -1,15 +1,23 @@
 package ecommerce.application
 
+import ecommerce.security.AuthInterceptor
 import ecommerce.security.LoginMemberArgumentResolver
 import org.springframework.context.annotation.Configuration
 import org.springframework.web.method.support.HandlerMethodArgumentResolver
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
 
 @Configuration
 class WebConfig(
     private val loginMemberArgumentResolver: LoginMemberArgumentResolver,
+    private val authInterceptor: AuthInterceptor,
 ) : WebMvcConfigurer {
     override fun addArgumentResolvers(resolvers: MutableList<HandlerMethodArgumentResolver>) {
         resolvers.add(loginMemberArgumentResolver)
+    }
+
+    override fun addInterceptors(registry: InterceptorRegistry) {
+        registry.addInterceptor(authInterceptor)
+            .addPathPatterns("/api/cart/**")
     }
 }
