+++ src/main/kotlin/ecommerce/security/LoginMemberArgumentResolver.kt
@@ -2,7 +2,6 @@ package ecommerce.security
 
 import ecommerce.annotation.LoginMember
 import ecommerce.exception.UnauthorizedException
-import ecommerce.service.AuthService
 import org.springframework.core.MethodParameter
 import org.springframework.stereotype.Component
 import org.springframework.web.bind.support.WebDataBinderFactory
@@ -11,10 +10,7 @@ import org.springframework.web.method.support.HandlerMethodArgumentResolver
 import org.springframework.web.method.support.ModelAndViewContainer
 
 @Component
-class LoginMemberArgumentResolver(
-    private val jwtTokenProvider: JwtTokenProvider,
-    private val authService: AuthService,
-) : HandlerMethodArgumentResolver {
+class LoginMemberArgumentResolver() : HandlerMethodArgumentResolver {
     override fun supportsParameter(parameter: MethodParameter): Boolean {
         return parameter.hasParameterAnnotation(LoginMember::class.java)
     }
@@ -25,21 +21,7 @@ class LoginMemberArgumentResolver(
         webRequest: NativeWebRequest,
         binderFactory: WebDataBinderFactory?,
     ): Any {
-        val token =
-            webRequest.getHeader("Authorization")
-                ?.removePrefix("Bearer ")
-                ?: throw UnauthorizedException("Authorization header is missing or malformed.")
-
-        if (!jwtTokenProvider.validateToken(token)) {
-            throw UnauthorizedException("Invalid or expired access token.")
-        }
-
-        val memberIdString = jwtTokenProvider.getSubjectFromToken(token)
-        val memberId =
-            memberIdString.toLongOrNull()
-                ?: throw UnauthorizedException("Invalid token subject (member ID).")
-
-        return authService.getMemberById(memberId)
+        return webRequest.getAttribute(AuthInterceptor.AUTHENTICATED_MEMBER, NativeWebRequest.SCOPE_REQUEST)
             ?: throw UnauthorizedException("Authenticated member not found in database.")
     }
 }
